refactor : Application 클래스에서 객체 분리
- InputCar, Race, Winner 를 Application에서 분리
